pyproject.toml
[project]
name = "every-eval-ever-pipeline"
version = "0.1.0"
description = "Middleware for ingesting and publishing AI evaluation results"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "psycopg[binary]>=3.1.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
]

[tool.setuptools.packages.find]
include = ["api*", "worker*", "shared*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
